refactor: simplify configuration writes and code cleanup
Simplified the routines for saving a configuration file. Add local function to generate nodes rather than individual creating each node.

Simplified the routines for loading and applying colors. A separate function was created to apply each color.

The remaining configuration loading can be simplified, but leaving as is for now.