feat: always popup notify regardless of history window
If popup notifications are enabled and a probe changes its status, always show the popup notification window. Previously, popup notifications would never display if the history window was open.

Also, change the order of routines when starting a probe to ensure the *start* history is recorded prior to performing the initial ping.