fix: don't show initial probe state in popup notification
Fixes an issue introduced in #782b236 where the initial probe state (up or down) is recorded to the status history. If the popup notification window happened to be visible when a new probe was recording its initial state, it incorrectly added this initial state to the popup notification.

Since the popup notification is intended to only display state changes, this fix ensures initial probe states aren't shown in the popup notification window.