feat: add high latency detection and alerting
New feature to alert on high latency for ICMP and TCP probes. When enabled, you can choose between two detection modes:

Fixed mode: Set a fixed threshold in milliseconds. An alert is triggered when a ping RTT is at or above this value.

Automatic mode: Set a threshold in milliseconds. An alert is triggered when a ping RTT exceeds the minimum observed RTT plus the threshold.

A high latency counter and alert threshold is used to determine when to alert. A high latency ping increments the counter (to a max of your threshold) and a low latency ping decrements the counter (to a minimum of zero). You are alerted of high latency if the counter hits your threshold. From high latency, you are alerted of normal latency if the counter drops back to zero.

**The ability to enable and configure this feature will be added in a separate commit.**

- Detection and alerting logic added to PerformIcmpProbe() when a ping response is received and to PerformTcpProbe() when a connection is made.
- Add `LatencyMode` enum to ApplicationOptions class to represent latency detection modes (off, auto, or fixed).
- Add `LatencyDetectionMode` to ApplicationOptions class to set which detection mode to use.
- Add `HighLatencyMilliseconds` to set the number of milliseconds used to consider if a ping is high latency.
- Add `HighLatencyAlertTiggerCount` to set the number of high latency pings needed to trigger an alert.
- Add `HighLatencyCount` to Probe class to track the number of high latency pings.
- Add `MinRtt` to Probe class to track the minimum observed RTT. Used when high latency detection is set for automatic mode (it's considered "high" when RTT is *x* milliseconds above the minimum observed RTT).